plugins/screens/map_screen/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "map-screen"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = ["pillow>=12.0.0"]

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points."vesselframe.plugins.screens"]
map_screen = "map_screen:make_plugin"

[project.entry-points."vesselframe.config.schemas"]
map_screen = "map_screen:get_config_schema"